requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
flask==3.1.0
groq==0.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
markupsafe==3.0.2
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
typing-extensions==4.12.2
urllib3==2.3.0
werkzeug==3.1.3
